options/internal,ansi: rename AllocatorLock to FutexLock and use frg::unique_lock
